# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.OverrideParticipationRequest:
    properties:
      date:
        type: string
      meal_type:
        type: string
      participating:
        type: boolean
      reason:
        type: string
      user_id:
        type: string
    required:
    - date
    - meal_type
    - participating
    - reason
    - user_id
    type: object
...
  handlers.SetParticipationRequest:
    properties:
      date:
        type: string
      meal_type:
        type: string
      participating:
        type: boolean
    required:
    - date
    - meal_type
    - participating
    type: object
  handlers.UpdatePreferencesRequest:
    properties:
      default_meal_preference:
        type: string
    required:
    - default_meal_preference
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@craftsbite.com
    name: API Support
  description: CraftsBite Backend API for meal management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CraftsBite API
  version: "1.0"
paths:
...
  /meals/all-teams-participation:
    get:
      description: Get today's meal participation for all teams (Admin/Logistics only)
      produces:
      - application/json
      responses:
        "200":
          description: All teams participation retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all teams participation
      tags:
      - meals
...
  /meals/participation:
    post:
      consumes:
      - application/json
      description: Set or update user's participation for a specific meal
      parameters:
      - description: Participation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetParticipationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Participation updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set meal participation
      tags:
      - meals
...
  /meals/participation/{date}:
    get:
      description: Get user's meal participation status for a specific date
      parameters:
      - description: Date (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Participation retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get participation by date
      tags:
      - meals
  /meals/participation/override:
    post:
      consumes:
      - application/json
      description: Admin or team lead can override user's meal participation
      parameters:
      - description: Override details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.OverrideParticipationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Participation overridden successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Override participation (Admin/Team Lead)
      tags:
      - meals
  /meals/team-participation:
    get:
      description: Get today's meal participation for team lead's team
      produces:
      - application/json
      responses:
        "200":
          description: Team participation retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get team participation
      tags:
      - meals
  /meals/today:
    get:
      description: Get tomorrow's meal schedule and user participation status
      produces:
      - application/json
      responses:
        "200":
          description: Meals retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get today's meals
      tags:
      - meals
...
  /users/me/preferences:
    get:
      description: Get current user's meal preferences
      produces:
      - application/json
      responses:
        "200":
          description: Preferences retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user preferences
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update current user's meal preferences
      parameters:
      - description: Preference details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Preferences updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user preferences
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"